feat: improve keep-alive timer management in VideoRtpSession
Added null check for rtp_session in keep_alive_timer_cb to prevent potential crashes. Enhanced setMuted method to manage the keep-alive timer based on mute state, ensuring proper timer setup and cancellation. Updated cancelKeepAliveTimer to reset the timer structure for better clarity.

+++ src/media/video/video_rtp_session.h
@@ -123,7 +123,7 @@ public:
     std::shared_ptr<VideoMixer> videoMixer_;
     std::shared_ptr<VideoInput> videoLocal_;
 
-    int getKaInterval() { return ka_inverval_; }
+    int64_t getKaInterval() { return ka_inverval_; }
     void startSender();
     void stopSender();
